refactor: Moved types and schematas to appropriate files.
