+++ app/components/theme-menu/theme-menu.tsx
@@ -1,7 +1,6 @@
 import cx from 'clsx';
-import type { ColorScheme } from '#types';
 import { includes } from '#utils/misc';
-import { useTheme } from '#utils/theme/theme.provider';
+import { type ColorScheme, useTheme } from '#utils/theme/theme.provider';
 import { Button, type Key, Menu, MenuItem, MenuItems } from '../(ui)';
 import { Icon, type IconName } from '../icon/icon';
