+++ app/utils/sync/championship.ts
@@ -1,6 +1,6 @@
-import type { ToastMessage } from '#types';
 import { db } from '#utils/db.server';
 import { invariant } from '#utils/misc';
+import type { ToastMessage } from '#utils/toast.server';
 import { getFirestoreChampionshipById } from '../firestore/championship';
 import { getLegacyMatches } from '../firestore/championship-match';
 import { getLegacyChampionshipPlayers } from '../firestore/championship-player';
